requirements.txt
# local package
-e .

# external requirements
Sphinx
coverage
awscli
flake8

auto_mix_prep==0.2.0
click==8.0.4
matplotlib==3.5.2
numpy==1.21.5
python-dotenv==0.21.0
scikit_learn==1.2.0
torch==1.13.1
